feat: optimize new page
