+++ src/app/(dashboard)/lookahead/page.tsx
@@ -1,4 +1,4 @@
-import LookaheadDashboard from "./_components/LookaheadDashboard";
+import LookaheadDashboard from "@/components/lookahead/LookaheadDashboard";
 
 export default function LookaheadPage() {
   return <LookaheadDashboard />;
